fix: change props, props type and storybook
